+++ src/Response/ServiceUnavailableResponse.php
@@ -1,0 +1,14 @@
+<?php
+namespace Kartenmacherei\RestFramework\Response;
+
+class ServiceUnavailableResponse extends ContentResponse
+{
+    /**
+     * @return int
+     */
+    protected function getResponseCode(): int
+    {
+        return 503;
+    }
+
+}